Issue #3029661 by Lendude, dawehner: Move BrowserTestBase tests out of modules/simpletest into /tests
(cherry picked from commit 185513febe85c73d3b6ae2939d2d32994d196bf9)
